tools/pan-photo-video/requirements.txt
moviepy>=1.0.3,<2
numpy>=1.24.0
Pillow>=10.0.0
